// src/app/_model/perfil-dto.ts
export interface PerfilDTO {
    idPozoInicio: number;
    nombrePozoInicio: string;
    alturaInicio: number;
    cotaTerrenoInicio: number;
    idPozoFin: number;
    nombrePozoFin: string;
    alturaFin: number;
    cotaTerrenoFin: number;
    pendiente: number;
    caudal: number;
    velocidad: number;
    diametro: number;
    material: string;
    longitud: number;
}